chore: move FormatCoins to x\/tx
